Continue implementing draggable gui actors
Add window base parenting. This means that all the actors in the
window have the same parent, a clone of a_gui that doesn't have
any other purpose than to serve as the base parent.

Having a common base parent makes it possible to move the whole
window around. It also allows transferring the parent status
to any other actor easily - just by making the new actor a parent
of the base parent actor. This will be useful for making it
possible to move the window by dragging the title bar.

The parenting system is not finished, and currently there's
no easy way to change the parent of the base parent.