Redo mouse operations
Game Editor's built-in mouse input system was
causing big problems when trying to implement
draggable gui elements, so changes were needed.

The new system works by using one single actor
for registering all the mouse inputs. This avoids
the problems of releasing a mouse button outside
of the actor it was pressed down on.