Fix a bug in actorExist and actorExist2
The functions rely on GE using the cloneindex -1 to indicate that
an actor is not a valid, existing actor. The actor count needs to
be checked because:

1. IF an actor with the given base name exists in the project
2. AND the lowest indexed clone (= the actor itself, if no clones present)
     in editor mode has "Create at startup" set to "No"
3. AND the actor currently has no alive instances in the game

..getclone will return an invalid Actor* that has the cloneindex of
that specific clone / actor over in editor mode despite it not existing
in the game. And that would break the functions without the ActorCount check.

Changes in the .ged project:
- remove a_guiHandler Key Down (enter) and Key Up (enter) events
    that were used at some point to change the ZDepths of the test windows
    for testing something relating the the ZDepths (duh)
- add the fixed misc.c to Global Code